fix admin file

+++ person_recognition/file_admin.py
@@ -17,14 +17,14 @@ def get_name_image():
 
 
 def deletFolder(path):
-    # Borrar archivos y subcarpetas
-    for root, dirs, files in os.walk(path, topdown=False):
-        for file in files:
-            os.remove(os.path.join(root, file))
-        for dir in dirs:
-            os.rmdir(os.path.join(root, dir))
-    # Borrar la carpeta principal
-    os.rmdir(path)
+    if os.path.exists(path):
+        for root, dirs, files in os.walk(path, topdown=False):
+            for file in files:
+                os.remove(os.path.join(root, file))
+            for dir in dirs:
+                os.rmdir(os.path.join(root, dir))
+        # Borrar la carpeta principal
+        os.rmdir(path)
 
 
 def newFolder(path):
